major changes
Major:
1. Fixed the Uniswap V3 Oracle bug. Updated OracleLibrary.
2. Added IUSDs as the interface of USDs
3. In VaultCore made some updates on rebase / allocate feature. Still in progress. Separated adding & updating a collateral. Changed updating a strategy to adding a strategy only.
4. Removed OracleGenesis. Now "default USDs price is 1 when no pool" is implemented in Oracle.

Minor:
1. Removed interface IBancorFormula
2. Changed all files to use OwnableUpgradeable in openzeppelin package instead of the local file
3. All address are now set up in initializer instead of hard-coded
4. Deleted unused files
5. Some minor updates inside VaultCoreLibrary

+++ contracts/interfaces/IVaultCore.sol
@@ -13,7 +13,7 @@ interface IVaultCore {
     function chi_alpha() external view returns (uint32);
     function chi_alpha_prec() external view returns (uint64);
     function chi_prec() external view returns (uint64);
-    function chiInit() external view returns (uint64);
+    function chiInit() external view returns (uint);
     function chi_beta() external view returns (uint32);
     function chi_beta_prec() external view returns (uint16);
     function chi_gamma() external view returns (uint32);
